travel_brain/requirements.txt
# ── Core HTTP & Parsing ───────────────────────────────────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
httpx==0.27.0

# ── YouTube ───────────────────────────────────────────────────────────────────
youtube-transcript-api==0.6.2
google-api-python-client==2.126.0

# ── Reddit ────────────────────────────────────────────────────────────────────
praw==7.7.1

# ── Text Processing ───────────────────────────────────────────────────────────
tiktoken==0.7.0
langdetect==1.0.9
ftfy==6.2.0
nltk==3.8.1
regex==2024.4.16

# ── Embeddings ────────────────────────────────────────────────────────────────
openai==1.30.1
sentence-transformers==3.0.1
torch==2.3.0

# ── Vector Databases ─────────────────────────────────────────────────────────
pinecone-client==3.2.2
chromadb==0.5.0

# ── Scheduling ────────────────────────────────────────────────────────────────
APScheduler==3.10.4

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tqdm==4.66.4
tenacity==8.3.0
rich==13.7.1
click==8.1.7
hashlib2==1.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0          # For testing FastAPI endpoints

# ── Phase 2: Query API + Chat ────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
google-generativeai>=0.8.0    # Gemini (free LLM for chat)
